+++ lib/rubocop/comment_config.rb
@@ -107,18 +107,14 @@ module RuboCop
       return if processed_source.comments.nil?
 
       processed_source.comments.each do |comment|
-        directive = directive_parts(comment)
-        next unless directive
-
-        yield comment, *directive
+        comment.text.scan(COMMENT_DIRECTIVE_REGEXP) do |match_captures|
+          yield comment, *directive_parts(match_captures)
+        end
       end
     end
 
-    def directive_parts(comment)
-      match = comment.text.match(COMMENT_DIRECTIVE_REGEXP)
-      return unless match
-
-      switch, cops_string = match.captures
+    def directive_parts(match_captures)
+      switch, cops_string = match_captures
 
       cop_names =
         cops_string == 'all' ? all_cop_names : cops_string.split(/,\s*/)